Add Animation Component
A new built in component, Animation, has been added. This component is
similar to Sprite - but it cycles through sequential sprites to provide
built in animation capability. Member variables frameLength and repeat
were added to control how long each sprite should be displayed and if
the animation should repeat respectively.