Add KeyMap inner class to Game.java
Inner class KeyMap has been added to Game.java. This class wraps around
a HashMap to map keycodes to a boolean. This change was made because
Game.java now requires a second map of keycodes to represent when keys
are released. Previously, Game.java only kept track of whether a key is
being pressed or not. This feature is useful as it allows for user to
check for single presses of a key instead of when a key is being held.