Move drawing methods to Canvas class
Drawing methods have been moved from Game to Canvas. Canvas abstracts
drawing to the screen therefore it makes more sense for these methods to
be located in this class. In addition to this, a method drawText has
been added to canvas which allows user to draw text to the canvas. All
built in components that draw to the canvas have been updated to reflect
this change.