Add Scene.addGameObject single group name overload
A new overloaded version of Scene.addGameObject has been added.
Previously adding a gameobject with a group required users to pass in a
list of strings. Creating a list of length 1 was very tedious for
objects with 1 group - therefore an alternate version of
Scene.addGameObject has been added that expects a single string.