Add support for groups to Collider.java
Colliders can now use groups to specify which types of objects it is
concerned colliding with. Collision behavior can now be specified by
overriding the onCollide method that expects a groupname to be passed
in. This change both allows for objects to collide differently with
specific groups and increases performance by preventing unecessary
collision checks.