Change Game.drawText coordinate system and drawing location
Game.drawText was drawing text by placing the texts bottom left hand
corner on the specified position. This has been changed to draw text
similarly to how sprites are currently drawn: centering the drawn object
on the passed in coordinates. The length of the text and the height of
the font is taken into account to center the text on the specified
point.

Game.drawText has also been changed to use the same coordinate system as
drawing sprites- with a (0,0) origin at the bottom left of the screen.