Change game window size to include window title bar and border
Game window creation has been changed to account for the window title bar
and border when creating the window. Previously Game would use the
passed in resolution for the size of the entire window. Instead Game
should have been using these dimensions only for the drawing area. A larger
window should have been created that would fit both the specified sized
drawing area and the title bar and borders.

This was accounted for previously by adding a constant 39 pixels to the vertical resolution of the screen.
However, this value was only estimated as the height of the title bar
and borders were not accounted for.

A more reliable way of accounting for these offsets has been
achieved by using ContentPane.setPreferredSize to create the window
dimensions. This method forces the area within the window borders to
have the correct user defined dimensions and the window will size up accordingly
to include the border and title bar.