Change Game.mousePosition to be relative to Canvas
Game.mousePosition was including the border and title bar in its
coordinates which should have been ignored. For example, when the mouse
was placed at the top left of the canvas, its position should be (0,0).
However, due to the window title bar, the mouse position would be (0 ,
TITLE BAR HEIGHT). Game.mouseMoved has been changed to create the former
behavior.

Resolves #13