Use `SceneManager` for scene transitions
Started updating AotB to to use the new scene manager. Code compiles but some components like `FollowPlayer` try to access groups before the scene is finished loading. This doesn't work because the scene manager defers adding new objects to scene and updating groups until after `Scene.load` completes.

Next step to fix this is to add a method for initiailization to `GameObject` and `Component`, then update `SceneManager` to call that method for game objects only when they are added to the scene.