Add constructor to Sprite that does not require an image path
Sprite now has a constructor that does not require an image path. This
will add a sprite component to the scene that will render nothing. This
is necessary as users may need to dynamically set the sprite image. A
constant has been added: `Sprite.DEFAULT_SPRITE`, that represents a 1x1 transparent image.