Remove Animation from RenderingSystem
`Animation` is no longer a renderable component, and instead is a
controller wrapper around `Sprite`. This change was made because
Animations are rendered the same way as a sprite - by getting its
parents transform information and the current sprite from the Animation.
Because of this, it is redundant for the `RenderingSystem` to render these components separately.

`Animation` now internally creates a new sprite component and adds it to
its parent. `Animation` will dynamically change `Sprite.sprite` to
contain the correct animation frame. Because this sprite has been added
to the parent game object, it will get caught by
`SceneManager.renderSprites`and be rendered to the screen.