Split Component class into Component and Script classes
Previously Component was an interface that had a single update method.

This design was problematic for the following reasons:

- Component did not represent what a component truly is: something that
  gets attached to a game object. Instead component was being used as
  something that should be updated every frame, which is instead a
  specific type of component.

- Components did not require a parent game object. This prevented
  systems from getting the parent game object from a Component
  reference. An example of this is seen in SceneManager.getComponents.
  Systems request a list of components from the SceneManager but need access to
  the parent game object to get its transform data. As
  a work around, SceneManager.getComponents returns the list of
  game objects that found matching components. The system would then
  need to call getComponents on each gameObject to get the desired
  component.

This has been redesigned by turning Component into a normal class and
creating a `Script` class that extends component and provides methods
for hooking into the update loop: update, fixedUpdate, start, and
onSceneLoad. The Component class provides a constructor that requires a
parent game object be passed in.