Change SceneManager.getComponents to return Components
SceneManager.getComponents was returning GameObjects because previously there was
no guaranteed way to get a parent game object from a Component
reference and systems need access to both.

To solve this, SceneManager.getComponents would return a list of
GameObjects so that systems could eventually call
GameObject.getComponents for the type they actually needed. However,
this was indirect and redundant as this operation was being completed
twice, once in SceneManager and a second time in the corresponding
system.

Now that Components all have a parent gameObject attribute,
SceneManager.GetComponents can return Components and systems can get
their parents from them directly.